Added simple implementation of predator prey in pred-prey-simple.py

+++ Assignment2.py
@@ -1,11 +1,5 @@
-import json
-from dataclasses import dataclass, field
 import random
-from statistics import fmean
-
-from polars import Boolean
-from pygame import Vector2
-from pygame.examples.moveit import WIDTH
+from dataclasses import dataclass, field
 from vi import Agent, Config, Simulation, Window, HeadlessSimulation
 from vi.util import count, probability
 
@@ -13,11 +7,11 @@ from vi.util import count, probability
 class PredatorPreyConfig(Config):
     width : int = 100
     height : int  = 100
-    radius : int = 25
+    radius : int = 20
     movement_speed : float = 5.0
     seed : int = 1
     duration : int = 5001
-    window : Window = field(default_factory=lambda: Window(width=800, height=800))
+    window : Window = field(default_factory=lambda: Window(width=1000, height=1000))
 
 @dataclass
 class PredatorConfig(Config):
@@ -25,7 +19,7 @@ class PredatorConfig(Config):
     height : int  = 100
     radius : int = 500
     movement_speed : float = 5.0
-    #seed : int = 1
+    seed : int = 1
     duration : int = 5001
 
 
@@ -34,11 +28,12 @@ class Predator(Agent[PredatorConfig]):
         super().__init__(*args, **kwargs)
         self.state = "wander"  # Initial state
         self.ticks = 0
-        self.energy = 120
-        self.ticks_to_reproduce = 250
-        self.reproduction_chance = 0.1  # Percentage chance to reproduce
-        self.energy_for_reproduction = 90  # Energy threshold for reproduction
-        self.hunt_chance = 0.5  # Percentage chance to successfully hunt prey
+        self.energy = 125
+        self.ticks_to_reproduce = 175
+        self.reproduction_chance = 1  # Percentage chance to reproduce
+        self.energy_for_reproduction = 80  # Energy threshold for reproduction
+        self.hunt_chance = 1  # Percentage chance to successfully hunt prey
+        self.speed = 0.7
 
     def update_ticks(self) -> None:
         self.ticks += 1
@@ -68,20 +63,18 @@ class Predator(Agent[PredatorConfig]):
 
 
     def hunt_prey(self) -> None:
-        if self.energy > 90:
-            return
         prey = self.detect_prey()
-        if prey:
-            if probability(self.hunt_chance):
-                prey.kill()
-                self.energy += 25 # Gain energy from eating prey
+
+        if prey and probability(self.hunt_chance):
+            prey.kill()
+            self.energy += 10 # Gain energy from eating prey
 
     def update(self) -> None:
-        self.pos += self.move
+
         self.hunt_prey()
         self.check_energy()
         self.update_ticks()
-
+        self.pos += self.move * self.speed
 
 
 class Prey(Agent[PredatorPreyConfig]):
@@ -89,8 +82,9 @@ class Prey(Agent[PredatorPreyConfig]):
         super().__init__(*args, **kwargs)
         self.state = "wander"  # Initial state
         self.ticks = 0
-        self.ticks_to_reproduce = 160
-        self.reproduction_chance = 0.7  # Percentage chance to reproduce
+        self.ticks_to_reproduce = random.randint(80, 100)
+        self.reproduction_chance = 0.9  # Percentage chance to reproduce
+        self.speed = 0.4
 
     def update_ticks(self) -> None:
         self.ticks += 1
@@ -98,23 +92,24 @@ class Prey(Agent[PredatorPreyConfig]):
     def update_breed(self) -> None:
         if probability(self.reproduction_chance):
             self.reproduce()
+            self.ticks_to_reproduce = random.randint(80, 100)
 
     def update(self) -> None:
-        self.pos += self.move
+
         if self.ticks == self.ticks_to_reproduce:
             self.update_breed()
             self.ticks = 0
 
         self.update_ticks()
-
+        self.pos += self.move * self.speed
 
 def run_sim():
-        (
-            Simulation(PredatorPreyConfig())
-            .batch_spawn_agents(50, Prey, images=["images/triangle.png"])
-            .batch_spawn_agents(2, Predator, images=["images/triangle@50px.png"])
-            .run()
-        )
+    (
+        Simulation(PredatorPreyConfig())
+        .batch_spawn_agents(100, Prey, images=["images/prey.png"])
+        .batch_spawn_agents(10, Predator, images=["images/predator.png"])
+        .run()
+    )
 
 
 if __name__ == "__main__":
